reset script

+++ supportBox/SBInfo/server/reset.php
@@ -3,22 +3,41 @@ require_once __DIR__."/../../../libraries/PhpFileDB.class.php";
 require_once __DIR__."/SBUtil.php";
 
 $C = SBUtil::dbConnection();
-$Q = $C->query("SHOW TABLES;");
-while($R = $Q->fetch_object())
-	$C->query("DROP TABLE ".$R->Tables_in_supportbox);
+if(mysqli_connect_error()){
+	echo mysqli_connect_error();
+	die();
+}
 
+echo "Datenbank-Verbindung aufgebaut…\n";
+
+$Q = $C->query("SHOW TABLES;");
+while($R = $Q->fetch_object()){
+	echo "Lösche Tabelle ".$R->Tables_in_supportbox.": ";
+	$C->query("DROP TABLE ".$R->Tables_in_supportbox);
+	if($C->error)
+		echo $C->error;
+	else
+		echo "OK";
+	echo "\n";
+}
 
 $aliases = "postmaster:    root
 root: pi
 ";
 
+echo "Setze Postfix alias zurück\n";
 exec("echo \"$aliases\" | sudo tee /etc/aliases > /dev/null");
 exec("sudo postalias /etc/aliases");
 
+echo "Entferne Keys\n";
 exec("sudo -u pi rm /home/pi/.ssh/id_rsa");
 exec("sudo -u pi rm /home/pi/.ssh/id_rsa.pub");
 
+echo "Setze Postfix relayhost zurück\n";
 exec("echo \"\" | sudo tee /etc/postfix/sasl_passwd  > /dev/null");
 exec("sudo postmap /etc/postfix/sasl_passwd");
 
+echo "Starte Dienste neu…\n";
 exec("sudo /usr/bin/supervisorctl restart all");
+
+echo "Zurücksetzen abgeschlossen!\n";
